AuditLogs: Better database output for computing diff
Audit logs will need to compute diff summary and have access to previous value when
updating something.
A snapshot of the previous value will be returned in an additional _old field, that will be
dropped before giving it to the web interface.
Also improving the returned value in case of error, because we will also log errors in audit
(for POST/PUT/DELETE).