[token] Authentication refactor
* Removing falcon-auth dependency, and getting more control on the JWT middleware
* Getting a typed authorization payload that is less nested

Some additional fixes (rebase):
1) Authorization policy None case
2) We're relying too much on 401

When we get the web interface for the first time, we're doing a lot of
different unauthenticated queries without a JWT token.
All these queries return 401.
We intercept all 401 to force the /login page.
It is more simple to intercept queries before they are sent, and get a
token if we don't have one.
That way, we can return 400 when there is a missing Authentication token
(which is the correct code).
We should also be careful, since /web can technically return 401, and
thus force us to the login page, we should maybe return a custom HTTP
error class for the following situations:
* Token expired (need to re-issue one)
* Wrong token (doesn't correspond to server private key)
* Not a JWT token
* etc

3) Authentication not required for these routes
4) Changes in auth payload
5) Handling CORS better
6) Context syntax correction
7) Auth payload syntax correction
8) Need fake token in cluster tests