[retry] Using tenacity for our base thread
Using a retry rate limit strategy for most of our threads.
The reasonning is as follow:
* We want threads to retry again if they encounter an error, to maximize
  availability.
* We want threads to fail if the retrying keeps happening (e.g. a configuration
error, system error, database error, etc). That way the administrator should be
informed of the error via monitoring.

Fixes (rebase):
1) Wrong argument
2) Better logs for threading

It was impossible to know when an error happened and we retryed, we
were losing the logs if it was just a one-off failure.
Let's log it, and keep the fact it's a retry.