[db] Optimization on comment housekeeping
In dev environment, the comment housekeeping, which was triggered at
startup, was stressing the database master (100% CPU usage), and was
not terminating, even after a few minutes. It affected the database
access, and increased the processing time for every alert received
(up to a few seconds).

The following was done to optimize the db call:
1) Group by record_uid before the lookup. Since the lookup is costly,
we want to perform the lookup once per unique record_uid, rather than
once per comment.
2) Batching the query. Instead of computing the list based on the lookup
at once, do it and perform the delete block by block. This way, even if
the job doesn't terminate, the work is not lost (it's executed partially
since we deleted elements already). This should also use less CPU of the
database, so even if the job takes a long time, it should have less
impact on the alert processing.