[config] Making configuration consider environment priority
Environment variables now properly override the yaml config, even
after the refresh of the configuration.

To achieve that, we needed to use pydantic.BaseSettings, but it ended
up being too limited in functionalities. We ended up taking inspiration
from pydantic.BaseSettings, so it can be reloaded, and config can be
flushed.

Also, to simplify, MetadataConfig no longer inherits from
ReadOnlyConfig, so we can simplify the __init__.

Also: Removing _class_get and _class_set
They are now unused, only _lock was using it