[config] Mistake in usage of FileLock
From: https://py-filelock.readthedocs.io/en/latest/index.html
Don’t use a FileLock to lock the file you want to write to, instead create a separate .lock file as shown above.

When an validation error happened, the Path.write_text was not called,
so the file became empty because of FileLock.