Supporting updates that verify all validators
1) Validators need to be verified all at once.
We're cloning the object with the new values, and we're using the
clone to update the current object's keys.

Also adding a test that verifies the update is working.

2) Excluded fields had a problems when updated via forms. By defaults,
forms were trying to set it to a falsy value ("" in javascript). To
handle this case, we will ignore updates of excluded fields when the
value provided is falsy.