[api] Fix the behavior of return_to
When we started returning a 400 (missing header) when no token was given,
we modified the behavior of the web interface, trying to go to the login
page when no token was present.

This caused an issue, since 2 different interceptors tried to insert
URLs with a return_to, we started getting nested return_to in the URL,
resulting in multiple clicks being necessary to enter the page.

This change is making sure we leave only one return_to argument. We
check before inserting.