Replacing the cluster with a syncer
We had many unreliable behavior with the clustering system.
Often the cluster will start getting out-of-sync easily.

As such, we will replace this with a thread dedicated to sync
in-memory values with the db.

To make it cheap, a value in the database will be updated when an
endpoint value is changed (POST/PUT/DELETE).

Hitting two birds with one stone, we will also maintain the current
sync state of each node as well. This will allow us to display to the
end user which nodes are in the cluster, and which nodes are in sync or
not, so we can monitor it if it ever happens.

Also, to make it more consistent and easier to debug, the syncing will
not be triggered on the local node where the change was performed,
making all nodes equal in term of sync, and also separating the API/db
changes from the processing plugins/syncer.
We often had different behavior that were hard to debug on the node
where the change happened vs the others, so let's make it easier this
time.