Refactor logging
Several points here:
* Loggers are separated into different files. The process logs
(generated for each new alerts) have their own file, and are rotated
differently (based on size). The API logs will be the logs generated
for each API call, and the main logs will be the one related to the
different threads (main thread starting, housekeeping, syncer, etc).
* The process and API logs will be able to be tracked with opentelemetry
traceID/spanID when we implement it.
* The logging configuration is greatly simplified, to leave only a few
sensible choices to the user (output, format and level)
* JSON log format is now supported
* Some logs were added/removed based on the new system