Adding support for opentelemetry traces
We tried several times to refactor the logs to be able to trace alerts
better when a unexpected behavior happened.
Traces are here to replace logging for tracing alerts. It works great
to get the context behind every processed log, but also every API call.
We will later update the other snooze plugins (in particular the ones
running as a daemon) to pass the traceID using W3C trace context.